refacto - creation section dans home Page
