+++ FRONTEND/src/app/components/home-page/home-page.component.html
@@ -1,0 +1,7 @@
+<div style="background-color: #f5f5f5; padding: 4rem 2rem;">
+  <app-section-introduction></app-section-introduction>
+  <app-section-a-propos></app-section-a-propos>
+  <app-section-engagements></app-section-engagements>
+</div>
+
+
