+++ FRONTEND/src/app/components/home-page/sections/section-a-propos/section-a-propos.component.html
@@ -1,0 +1,26 @@
+<section style="margin-block: 6rem;">
+<div style="text-align: center; margin-bottom: 3rem;">
+    <h2 style="font-size: 2rem; font-weight: bold; margin: 0;">À propos de nous</h2>
+  </div>
+
+  <div style="display: flex; flex-wrap: wrap; align-items: center; justify-content: center; max-width: 1200px; margin: auto; margin-bottom: 5rem;">
+    <!-- Image -->
+    <div style="flex: 1; min-width: 300px; padding: 1rem;">
+      <img src="assets/images/home/engagement.png" alt="Engagement social" style="width: 100%; height: auto; border-radius: 8px;" />
+    </div>
+
+    <!-- Texte -->
+    <div style="flex: 1; min-width: 300px; padding: 1rem;">
+      <h3 style="font-size: 1.5rem; font-weight: bold; margin-bottom: 1rem;">Nos valeurs</h3>
+      <p style="color: #555; line-height: 1.6;">
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eu turpis molestie, dictum est a, mattis tellus.
+        Sed dignissim, metus nec fringilla accumsan, risus sem sollicitudin lacus, ut interdum tellus elit sed risus.
+        Maecenas eget condimentum velit, sit amet feugiat lectus. Class aptent taciti sociosqu ad litora torquent per
+        conubia nostra, per inceptos himenaeos. Praesent auctor purus luctus enim egestas, ac scelerisque ante pulvinar.
+      </p>
+      <button style="margin-top: 1rem; padding: 0.5rem 1rem; background-color: #333; color: white; border: none; border-radius: 4px; cursor: pointer;">
+        Show more
+      </button>
+    </div>
+  </div>
+</section>
