+++ FRONTEND/src/app/components/home-page/sections/section-a-propos/section-a-propos.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-section-a-propos',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './section-a-propos.component.html'
+})
+export class SectionAProposComponent {}
