+++ FRONTEND/src/app/components/home-page/sections/section-engagements/section-engagements.component.html
@@ -1,0 +1,25 @@
+<!-- Section Nos engagements -->
+<section style="margin-block: 6rem; text-align: center;">
+    <h2 style="font-size: 2rem; font-weight: bold; margin-bottom: 3rem;">Nos engagements</h2>
+    <div style="display: flex; flex-wrap: wrap; justify-content: center; gap: 3rem;">
+
+        <!-- Engagement 1 : Écologie -->
+        <div style="width: 200px;">
+        <img src="assets/icons/eco.png" alt="Écologie" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
+        <p style="font-weight: bold;">Empreinte écologique basse</p>
+        </div>
+
+        <!-- Engagement 2 : Associations -->
+        <div style="width: 200px;">
+        <img src="assets/icons/donation.png" alt="Associations" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
+        <p style="font-weight: bold;">Reversement à des associations</p>
+        </div>
+
+        <!-- Engagement 3 : Transparence -->
+        <div style="width: 200px;">
+        <img src="assets/icons/transparency.png" alt="Transparence" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
+        <p style="font-weight: bold;">Transparence totale</p>
+        </div>
+    </div>
+</section>
+
