+++ FRONTEND/src/app/components/home-page/sections/section-engagements/section-engagements.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-section-engagements',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './section-engagements.component.html'
+})
+export class SectionEngagementsComponent {}
