+++ FRONTEND/src/app/components/home-page/sections/section-introduction/section-introduction.component.html
@@ -1,0 +1,22 @@
+<div style="display: flex; justify-content: center; align-items: center;">
+    <div style="display: flex; max-width: 1200px; width: 100%; padding: 2rem;">
+      
+      <!-- Bloc texte -->
+      <div style="flex: 1; padding: 1rem;">
+        <h1 style="font-size: 2.5rem; margin-bottom: 1rem;">Les premiers maillots<br>avec un impact social</h1>
+        <p style="font-size: 1rem; margin-bottom: 2rem; color: #555;">
+          Découvrez nos maillots éthiques : production européenne, matériaux recyclables<br>
+          et reversement d’une partie des bénéfices à des associations.
+        </p>
+        <button style="padding: 0.75rem 1.5rem; background-color: #4f46e5; color: white; border: none; border-radius: 4px; cursor: pointer;">
+          Voir nos collections
+        </button>
+      </div>
+  
+      <!-- Bloc image -->
+      <div style="flex: 1; display: flex; justify-content: center; align-items: center;">
+        <img src="/assets/images/home/home.png" alt="Maillot UrbanHeritage" style="width: 100%; height: auto; border-radius: 8px;" />
+      </div>
+  
+    </div>
+  </div>
