+++ FRONTEND/src/app/components/home-page/sections/section-introduction/section-introduction.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-section-introduction',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './section-introduction.component.html'
+})
+export class SectionIntroductionComponent {}
