--- FRONTEND/src/app/components/homePage/home-page.component.html
@@ -1,82 +1,0 @@
-<div style="display: flex; justify-content: center; align-items: center; min-height: 100vh; background-color: #f5f5f5;">
-  <div style="display: flex; max-width: 1200px; width: 100%; padding: 2rem;">
-    
-    <!-- Bloc texte -->
-    <div style="flex: 1; padding: 1rem;">
-      <h1 style="font-size: 2.5rem; margin-bottom: 1rem;">Les premiers maillots<br>avec un impact social</h1>
-      <p style="font-size: 1rem; margin-bottom: 2rem; color: #555;">
-        Découvrez nos maillots éthiques : production européenne, matériaux recyclables<br>
-        et reversement d’une partie des bénéfices à des associations.
-      </p>
-      <button style="padding: 0.75rem 1.5rem; background-color: #4f46e5; color: white; border: none; border-radius: 4px; cursor: pointer;">
-        Voir nos collections
-      </button>
-    </div>
-
-    <!-- Bloc image -->
-    <div style="flex: 1; display: flex; justify-content: center; align-items: center;">
-      <img src="/assets/images/home/home.png" alt="Maillot UrbanHeritage" style="width: 100%; height: auto; border-radius: 8px;" />
-    </div>
-
-  </div>
-</div>
-
-<!-- Section À propos -->
-<section style="background-color: #f5f5f5; padding: 2rem 2rem;">
-  <div style="text-align: center; margin-bottom: 3rem;">
-    <h2 style="font-size: 2rem; font-weight: bold; margin: 0;">À propos de nous</h2>
-    <p style="color: #666;">Présentation des valeurs de l’entreprise</p>
-  </div>
-
-  <div style="display: flex; flex-wrap: wrap; align-items: center; justify-content: center; max-width: 1200px; margin: auto;">
-    <!-- Image -->
-    <div style="flex: 1; min-width: 300px; padding: 1rem;">
-      <img src="assets/images/home/engagement.png" alt="Engagement social" style="width: 100%; height: auto; border-radius: 8px;" />
-    </div>
-
-    <!-- Texte -->
-    <div style="flex: 1; min-width: 300px; padding: 1rem;">
-      <h3 style="font-size: 1.5rem; font-weight: bold; margin-bottom: 1rem;">Nos valeurs</h3>
-      <p style="color: #555; line-height: 1.6;">
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eu turpis molestie, dictum est a, mattis tellus.
-        Sed dignissim, metus nec fringilla accumsan, risus sem sollicitudin lacus, ut interdum tellus elit sed risus.
-        Maecenas eget condimentum velit, sit amet feugiat lectus. Class aptent taciti sociosqu ad litora torquent per
-        conubia nostra, per inceptos himenaeos. Praesent auctor purus luctus enim egestas, ac scelerisque ante pulvinar.
-      </p>
-      <button style="margin-top: 1rem; padding: 0.5rem 1rem; background-color: #333; color: white; border: none; border-radius: 4px; cursor: pointer;">
-        Show more
-      </button>
-    </div>
-  </div>
-</section>
-
-
-<!-- Section Nos engagements -->
-<section style="background-color: #f5f5f5; padding: 4rem 2rem; text-align: center;">
-  <h2 style="font-size: 2rem; font-weight: bold; margin-bottom: 3rem;">Nos engagements</h2>
-
-  <div style="display: flex; flex-wrap: wrap; justify-content: center; gap: 3rem;">
-    
-    <!-- Engagement 1 : Écologie -->
-    <div style="width: 200px;">
-      <img src="assets/icons/eco.png" alt="Écologie" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
-      <p style="font-weight: bold;">Empreinte écologique basse</p>
-    </div>
-
-    <!-- Engagement 2 : Associations -->
-    <div style="width: 200px;">
-      <img src="assets/icons/donation.png" alt="Associations" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
-      <p style="font-weight: bold;">Reversement à des associations</p>
-    </div>
-
-    <!-- Engagement 3 : Transparence -->
-    <div style="width: 200px;">
-      <img src="assets/icons/transparency.png" alt="Transparence" style="width: 50px; height: 50px; margin-bottom: 1rem;" />
-      <p style="font-weight: bold;">Transparence totale</p>
-    </div>
-
-  </div>
-</section>
-
-
-
