--- FRONTEND/src/app/components/homePage/home-page.component.ts
@@ -1,11 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { Component } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-@Component({
-  selector: 'app-home-page',
-  standalone: true,
-  imports: [CommonModule, RouterModule],
-  templateUrl: './home-page.component.html'
-})
-export class HomePageComponent {}
